Draft - Include more information on AlreadySentError errors
We have an app that is getting a lot of AlreadySentErrors. Likely
somewhere we have missed an `halt()`, but tracking that down is tricky.
I wonder if we could supply some more information in the various places
the error gets raised to help someone debug the issue?

Exact wording and contents can be discussed.